+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Fixed
 - Select `all` by default in profile viewer ([#3968](https://github.com/julia-vscode/julia-vscode/pull/3968))
+- Also show "Copy Plot" button for HTML plots ([#3969](https://github.com/julia-vscode/julia-vscode/pull/3969))
 
 ## [1.171.0] - 2026-01-06
 ### Fixed